fix for sql server

+++ GlobusObservability.Core/Entities/MetricModel.cs
@@ -9,6 +9,6 @@ namespace GlobusObservability.Core.Entities
         
         public string Duration { get; set; }
 
-        public List<Dictionary<string, Dictionary<string, int[]>>> Value { get; set; }
+        public List<Dictionary<string, Dictionary<string, long[]>>> Value { get; set; }
     }
 }
